move lib outside angular src
