+++ packages/angular-playground/index.ts
@@ -1,5 +1,5 @@
 export * from './build/lib/api';
 export * from './build/lib/playground';
-export * from './build/playground.module';
-export * from './build/playground-common.module';
-export * from './build/app.component';
+export * from './build/src/playground.module';
+export * from './build/src/playground-common.module';
+export * from './build/src/app.component';
